Removed proxy information, don't delete blobs
The registry has a garbage-collect command that removes blobs
safely and reliably. Deleting blobs carelessly can be very bad
as layers are shared between images so extra checks must be done
to ensure used layers are kept. This application need only delete
the manifest. The registry will do the rest.